[solver] Add a generic edge index function
Eventually `generic_edge_index` could be re-written to use a const
generic parameter and take an array instead of a slice.

Currently `generic_edge_index` returns a u32, but should be changed to
either return a larger integer or a generic type that can be of a
variable sized integer.